Refactor move actions to 'gui.actions' namespace
