+++ securedrop_client/locale/messages.pot
@@ -59,6 +59,15 @@ msgstr ""
 msgid "Failed to delete source at server"
 msgstr ""
 
+msgid "All Files"
+msgstr ""
+
+msgid "Entire source account"
+msgstr ""
+
+msgid "Files and messages"
+msgstr ""
+
 msgid "SecureDrop Client {}"
 msgstr ""
 
@@ -227,15 +236,6 @@ msgid " to compose or send a reply"
 msgstr ""
 
 msgid "DELETE"
-msgstr ""
-
-msgid "All Files"
-msgstr ""
-
-msgid "Entire source account"
-msgstr ""
-
-msgid "Files and messages"
 msgstr ""
 
 msgid "Username"
